+++ src/utils/timeParser.ts
@@ -1,12 +1,12 @@
-const limit = 7 * 86400 * 1000;
-const pattern = /[0-9]+(?=s|m|h|d)/;
+// 7 * 86400 * 1000
+const limit: number = 604800000;
 
-export const parseTime = (time: string): boolean | number => {
-    if(!pattern.test(time))
+export default (time: string): false | number => {
+    if(!/[0-9]+(?=s|m|h|d)/g.test(time))
         return false;
 
-    const parsed = time.split(/[0-9]+(?=s|m|h|d)/);
-    let parsedTime = parseInt(parsed[0]);
+    const parsed: string[] = time.match(/[0-9]+|(s|m|h|d)/g);
+    let parsedTime: number = parseInt(parsed[0]);
 
     switch(parsed[1]) {
         case 's': parsedTime = parsedTime * 1000; break;
@@ -15,5 +15,5 @@ export const parseTime = (time: string): boolean | number => {
         case 'd': parsedTime = parsedTime * 86400000; break;
     }
 
-    return (parsedTime > limit && limit != -1) ? false : parsedTime;
+    return (parsedTime > limit && limit != -1 || parsedTime < 10000) ? false : parsedTime;
 }